Extract AbstractWritableNode
